+++ tests/src/Fixtures/DateTime/Constants.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Tests\Fixtures\DateTime;
+
+final readonly class Constants
+{
+    public const SOURCE_DATETIME = '2024-01-01 12:00:00';
+    public const SOURCE_DATETIME_EPOCH = 1704110400;
+}