+++ tests/src/Fixtures/DateTime/ObjectWithIntYYYYMMDD.php
@@ -13,14 +13,17 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\Tests\Fixtures\DateTime;
 
-final class ObjectWithDateTime implements DateTimeTestObjectInterface
+use Rekalogika\Mapper\Attribute\DateTimeOptions;
+
+final class ObjectWithIntYYYYMMDD implements DateTimeTestObjectInterface
 {
-    public ?\DateTime $property = null;
+    #[DateTimeOptions(format: 'Ymd')]
+    public ?int $property = null;
 
     public static function preinitialized(): static
     {
         $object = new static();
-        $object->property = new \DateTime(Constants::SOURCE_DATETIME);
+        $object->property = 20220105;
 
         return $object;
     }