+++ tests/src/Fixtures/DateTime/ObjectWithStringDDMMYYYYWithTimeZone.php
@@ -13,14 +13,17 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\Tests\Fixtures\DateTime;
 
-final class ObjectWithDateTime implements DateTimeTestObjectInterface
+use Rekalogika\Mapper\Attribute\DateTimeOptions;
+
+final class ObjectWithStringDDMMYYYYWithTimeZone implements DateTimeTestObjectInterface
 {
-    public ?\DateTime $property = null;
+    #[DateTimeOptions(format: 'd-m-Y', timeZone: 'Asia/Jakarta')]
+    public ?string $property = null;
 
     public static function preinitialized(): static
     {
         $object = new static();
-        $object->property = new \DateTime(Constants::SOURCE_DATETIME);
+        $object->property = '01-01-2023';
 
         return $object;
     }