+++ tests/src/Fixtures/DateTime/ObjectWithStringWithTimeZone.php
@@ -13,14 +13,17 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\Tests\Fixtures\DateTime;
 
-final class ObjectWithDateTime implements DateTimeTestObjectInterface
+use Rekalogika\Mapper\Attribute\DateTimeOptions;
+
+final class ObjectWithStringWithTimeZone implements DateTimeTestObjectInterface
 {
-    public ?\DateTime $property = null;
+    #[DateTimeOptions(timeZone: 'Asia/Jakarta')]
+    public ?string $property = null;
 
     public static function preinitialized(): static
     {
         $object = new static();
-        $object->property = new \DateTime(Constants::SOURCE_DATETIME);
+        $object->property = Constants::SOURCE_DATETIME;
 
         return $object;
     }