+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * test: attribute with missing class
 * fix: multiple attributes found in inheritance chain now correctly handled
 * feat: save all attributes in metadata, not just our attributes
+* feat: datetime transformation to int & float
 
 ## 1.8.0